Various fixes
